✨Comp-backend override resources with selected hardware (#4954)
