+++ packages/service-library/src/servicelib/rabbitmq/_constants.py
@@ -1,3 +1,4 @@
 from typing import Final
 
 BIND_TO_ALL_TOPICS: Final[str] = "#"
+RPC_REMOTE_METHOD_TIMEOUT_S: Final[int] = 30
