+++ services/clusters-keeper/tests/unit/test_rpc_ec2_instances.py
@@ -4,9 +4,11 @@
 
 import pytest
 from fastapi import FastAPI
-from models_library.api_schemas_clusters_keeper import CLUSTERS_KEEPER_RPC_NAMESPACE
 from models_library.api_schemas_clusters_keeper.ec2_instances import EC2InstanceType
-from servicelib.rabbitmq import RabbitMQRPCClient, RPCMethodName
+from servicelib.rabbitmq import RabbitMQRPCClient, RPCServerError
+from servicelib.rabbitmq.rpc_interfaces.clusters_keeper.ec2_instances import (
+    get_instance_type_details,
+)
 
 pytest_simcore_core_services_selection = [
     "rabbit",
@@ -34,10 +36,9 @@ async def test_get_instance_type_details_all_options(
     clusters_keeper_rabbitmq_rpc_client: RabbitMQRPCClient,
 ):
     # an empty set returns all options
-    rpc_response = await clusters_keeper_rabbitmq_rpc_client.request(
-        CLUSTERS_KEEPER_RPC_NAMESPACE,
-        RPCMethodName("get_instance_type_details"),
-        instance_type_names=set(),
+
+    rpc_response = await get_instance_type_details(
+        clusters_keeper_rabbitmq_rpc_client, instance_type_names=[]
     )
     assert rpc_response
     assert isinstance(rpc_response, list)
@@ -48,14 +49,23 @@ async def test_get_instance_type_details_specific_type_names(
     _base_configuration: None,
     clusters_keeper_rabbitmq_rpc_client: RabbitMQRPCClient,
 ):
-    # an empty set returns all options
-    rpc_response = await clusters_keeper_rabbitmq_rpc_client.request(
-        CLUSTERS_KEEPER_RPC_NAMESPACE,
-        RPCMethodName("get_instance_type_details"),
+    rpc_response = await get_instance_type_details(
+        clusters_keeper_rabbitmq_rpc_client,
         instance_type_names={"t2.micro", "g4dn.xlarge"},
     )
     assert rpc_response
     assert isinstance(rpc_response, list)
     assert len(rpc_response) == 2
-    assert rpc_response[0].name == "g4dn.xlarge"
     assert rpc_response[1].name == "t2.micro"
+    assert rpc_response[0].name == "g4dn.xlarge"
+
+
+async def test_get_instance_type_details_with_invalid_type_names(
+    _base_configuration: None,
+    clusters_keeper_rabbitmq_rpc_client: RabbitMQRPCClient,
+):
+    with pytest.raises(RPCServerError):
+        await get_instance_type_details(
+            clusters_keeper_rabbitmq_rpc_client,
+            instance_type_names={"t2.micro", "g4dn.xlarge", "invalid.name"},
+        )
