+++ services/director-v2/src/simcore_service_director_v2/modules/catalog.py
@@ -1,19 +1,19 @@
 import logging
 import urllib.parse
 from dataclasses import dataclass
-from typing import Any, TypeAlias
+from typing import Any
 
 import httpx
 from fastapi import FastAPI, HTTPException, status
 from models_library.services import ServiceKey, ServiceVersion
+from models_library.services_resources import ServiceResourcesDict
 from models_library.users import UserID
+from pydantic import parse_obj_as
 from settings_library.catalog import CatalogSettings
 
 from ..utils.client_decorators import handle_errors, handle_retry
 
 logger = logging.getLogger(__name__)
-
-ServiceResourcesDict: TypeAlias = dict[str, Any]
 
 
 def setup(app: FastAPI, settings: CatalogSettings) -> None:
@@ -90,7 +90,9 @@ class CatalogClient:
         )
         resp.raise_for_status()
         if resp.status_code == status.HTTP_200_OK:
-            json_response: ServiceResourcesDict = resp.json()
+            json_response: ServiceResourcesDict = parse_obj_as(
+                ServiceResourcesDict, resp.json()
+            )
             return json_response
         raise HTTPException(status_code=resp.status_code, detail=resp.content)
