+++ services/director-v2/src/simcore_service_director_v2/modules/db/repositories/comp_tasks/__init__.py
@@ -1,0 +1,3 @@
+from ._core import CompTasksRepository
+
+__all__: tuple[str, ...] = ("CompTasksRepository",)
